Condition interface added. It represents a condition that can be used to filer the partition generated by decamon.
FactoryConditions class added.
It is a factory class which generates possible conditions that can be
used by user and that can be passed to the distribution method in the
traceexpression class.

TraceExpression class modified to support the distribution with
corresponding generation of all possible minimal monitoring safe
partition usign decamon.